pyproject.toml
[project]
name = "card-inventory"
version = "0.1.0"
description = "Trading card inventory management system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[tool.setuptools]
packages = ["backend"]
